workflow: Remove the signing step
Having it automatically sign the web-ext build artifacts on a tagged
version automatically makes that version available as a custom published
version but it is not available in AMO (addons.mozilla.org) and is not
possible to republish the same version, thus effectively making it
private.

The preferred approach is to manually upload the specific artifact to
AMO and it automatically does the signing using my account.

Signed-off-by: Antonio Gutierrez <chibby0ne@gmail.com>